requirements.txt
numpy
pandas
setuptools
mysql-connector-python
scikit-learn
lightgbm
xgboost
tensorboard
torch
matplotlib
mlflow
dvc
flask
pylint
flake8
black